refactor: extract kanban session start workflows
